+++ release_notes.md
@@ -3,3 +3,4 @@
 <!-- Please add your release notes in the following format:
 - My change description (#PR)
 -->
+- Disable Diagnostic Events when Table Storage is not accessible (#10996)